Fixed the brightness options to work with mouse.
It is the title says.
However, there is one slight imperfection.
There is a code that auto sets the global.background_brightness_menu to 0 if within a range of -0.1 to 0.1.
I made it prefer black text color but this makes the text color change to white only when the slider is below -10%, you should be able to fix this by setting that auto-correction code to a release mouse event instead but I decided against messing with that at the moment.

It's just visual and since the slider jumps to 0 as soon as you let go within that range anyway it doesn't matter for the end user. It's simply a visual imperfection. I leave the decision up to Jonathan Nilsson whether or not to fix this.