Fixed with save files more
Fixed with save files more, now there is a save_files folder, and save files will be copied over to that folder. There seems to be a problem where the next level still wont unlock sometimes, but you can reset the game (go to title and then back) and it fixes itself, but if anyone get this problem still, then I know it's not just for me, I'm manipulating the save files so maybe it's something with that or not.