Many fixes
Resource pack now defaults sprites to default resource pack, if there is no sprites in current resource pack.
You can press Q or W to scroll objects in level editor now, instead of just scroll wheel being available.
Clicking on a object that has multiple forms now changes it, but you can hover mouse over object and press A if you want to.
"Open custom level folder" button in level editor pause menu.
Fixed so cardboard boxes doesn't float in the air if you zoom the camera in.
When generating a level map, it takes a little longer for it to generate, but it looks a bit more clear now, so things don't look squished because they just spawned in.