Added notification when you can't use 2nd controller in pause menu
When trying to connect different controllers on Steam Deck, it somehow got recognized as player 2 instead of player 1, and it confused us. We thought the game had crashed on the pause screen when we couldn't control the menu anymore. So there should be some sort of notification when you can't use a controller.